requirements/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.4
Pillow==10.1.0
